refactor: #15 ダブルクリックで編集できるフォームのコンポーネントを切り出し
